backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
pyportfolioopt==1.5.5
cvxpy==1.4.1
yfinance==0.2.33
python-dotenv==1.0.0
# New dependencies
supabase==2.3.0
psycopg2-binary==2.9.9